refactor: centralize template management and optimize Jinja template utilities
* Create centralized TemplateManager class to eliminate global template state
  - Replace global chat_template_name/completion_template_name variables
  - Integrate TemplateManager into _GlobalState and inject into serving classes
  - Add proper type hints for TokenizerManager and TemplateManager parameters

* Extract and reorganize Jinja template utilities for better separation of concerns
  - Create jinja_template_utils.py with template detection and processing logic
  - Move utilities from openai/utils.py to avoid improper dependencies
  - Rename detect_template_content_format -> detect_jinja_template_content_format

* Optimize template content format detection for better performance
  - Detect format during template loading instead of on every request
  - Cache detected format in TemplateManager.jinja_template_content_format property
  - Add logging for template format detection results

* Clean up codebase and improve maintainability
  - Remove unused imports and clean up import organization
  - Simplify TemplateManager interface by removing unused is_initialized property
  - Update all serving classes (chat, completions, embedding) to use dependency injection
  - Improve code organization and eliminate architectural debt

Benefits:
- ✅ Eliminates global state pollution
- ✅ Better separation of concerns (generic vs OpenAI-specific utilities)
- ✅ Improved performance through caching
- ✅ Cleaner dependency injection pattern
- ✅ More testable and maintainable architecture
